spawn_kernel_thread() now matches the spawn_thread() prototype - there is no reason for them to be different.
git-svn-id: file:///srv/svn/repos/haiku/trunk/current@7909 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ headers/os/drivers/KernelExport.h
@@ -91,7 +91,7 @@ extern bool			cancel_timer(timer *t);
 /* kernel threads */
 
 extern thread_id	spawn_kernel_thread(thread_func function, const char *threadName, 
-						long priority, void *arg);
+						int32 priority, void *arg);
 
 
 /*-------------------------------------------------------------*/
